+++ app/server/src/main/java/com/bsuir/tracker/controller/EmployeeController.java
@@ -3,7 +3,9 @@ package com.bsuir.tracker.controller;
 import java.io.IOException;
 import java.util.List;
 import javax.servlet.http.HttpServletRequest;
+import javax.xml.crypto.Data;
 
+import com.bsuir.tracker.controller.security.GetTokenService;
 import org.hibernate.exception.ConstraintViolationException;
 import org.jboss.logging.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,6 +21,7 @@ import com.bsuir.tracker.entity.EmployeeEntity;
 import com.bsuir.tracker.Service.EmployeeService;
 
 @Controller
+@RequestMapping(value = "/Backdoor")
 public class EmployeeController {
 
     private static  final Logger logger = Logger.getLogger(ImageController.class);
@@ -29,6 +32,8 @@ public class EmployeeController {
 
     @Autowired
     private  EmployeeService employeeService;
+    @Autowired
+    private GetTokenService getTokenService;
 
     @RequestMapping(value = "/Employees")
     public ModelAndView listEmployee(ModelAndView model) throws IOException {
@@ -49,35 +54,42 @@ public class EmployeeController {
     @RequestMapping(value = "/saveEmployee", method = RequestMethod.POST)
     public ModelAndView saveEmployee(@ModelAttribute("employee") @Validated EmployeeEntity employee, BindingResult bindingResult) {
         if(bindingResult.hasErrors()){
-            ModelAndView modelAndView = new ModelAndView("redirect:/errorView");
-            modelAndView.addObject("message", "Whoops, something gone wrong with your input!");
-            return modelAndView;
+            return GetErrorView("Whoops, something gone wrong with your input!");
         }
         try {
             if (employee.getIdemployee() == 0) {
                 employeeService.addEmployee(employee);
             } else {
-                employeeService.updateEmployee(employee);
+                if(employeeService.updateEmployee(employee) == null)
+                {
+                    return GetErrorView("Whoops, cannot change last admin!");
+                }
             }
         }
         catch (ConstraintViolationException e) {
-            ModelAndView modelAndView = new ModelAndView("redirect:/errorView");
-            modelAndView.addObject("message", "Whoops, something gone wrong with ID-s!");
-            return modelAndView;
+            return GetErrorView("Whoops, something gone wrong with ID-s!");
         }
         catch (DataIntegrityViolationException e) {
-            ModelAndView modelAndView = new ModelAndView("redirect:/errorView");
-            modelAndView.addObject("message", "Whoops, something gone wrong with SQL data integrity!");
-            return modelAndView;
+            return GetErrorView("Whoops, something can gone wrong with SQL data integrity!");
         }
-        return new ModelAndView("redirect:/Employees");
+        return new ModelAndView("redirect:/Backdoor/Employees");
     }
 
     @RequestMapping(value = "/deleteEmployee", method = RequestMethod.GET)
     public ModelAndView deleteEmployee(HttpServletRequest request) {
         int idEmployee = Integer.parseInt(request.getParameter("id"));
-        employeeService.deleteEmployee(idEmployee);
-        return new ModelAndView("redirect:/Employees");
+        try
+        {
+            if (idEmployee == getTokenService.getIdFromToken((String) request.getSession().getAttribute("token"))) {
+                return GetErrorView("Cannot delete yourself!");
+            }
+            employeeService.deleteEmployee(idEmployee);
+        }
+        catch (DataIntegrityViolationException e)
+        {
+            return GetErrorView("Whoops, something gone wrong with SQL data integrity!");
+        }
+        return new ModelAndView("redirect:/Backdoor/Employees");
     }
 
     @RequestMapping(value = "/editEmployee", method = RequestMethod.GET)
@@ -89,4 +101,9 @@ public class EmployeeController {
         return model;
     }
 
+    private ModelAndView GetErrorView(String message)
+    {
+        ModelAndView modelAndView = new ModelAndView("redirect:/errorView?message=" + message);
+        return modelAndView;
+    }
 }
